refactor(bridge): extract stream loop shell
#4230 S6: move the turn_bridge stream receive/drain loop shell, cancel gates, remaining stream event arms, long-running placeholder open/retarget state, runtime handoff delegation calls, stream/status tick call, and bridge latency span emission into turn_bridge/stream_loop.rs.

Control-flow mapping: external return/break/continue = 0/0/0 because the main loop moved intact; internal labelled break 'outer = 6, unlabelled internal break = 3, internal continue = 1. No return/?/.await? exits.

Threaded locals: StreamLoopState carries 60 mutable state fields; pending long-running open/retarget state is owned inside stream_loop.rs and returned through StreamLoopOutput.

Size gates: turn_bridge/mod.rs raw/prod 1690/1230; stream_loop.rs raw/prod 1638/1638. stream_loop.rs is registered as a #4230 discord-relay giant with deadline 2026-08-31 and admitted in the namespace cap config.

Gates: python3 scripts/generate_inventory_docs.py; python3 scripts/check_agent_maintenance_docs.py; bash scripts/ci-script-checks.sh; cargo fmt.
