+++ scripts/giant_file_registry.toml
@@ -507,6 +507,17 @@ deadline = "2026-08-31"
 decompose_issue = "#4230"
 
 [[entry]]
+# Created by #4230 S6 as the approved stream-loop shell extraction from
+# turn_bridge/mod.rs. It remains over the giant threshold because the event arms
+# share placeholder state, background-child tracking, transcript events, inflight
+# persistence, cancel gates, and stream-tick state; the approved plan marks this
+# loop shell irreducible for this stage.
+file = "src/services/discord/turn_bridge/stream_loop.rs"
+owner = "discord-relay"
+deadline = "2026-08-31"
+decompose_issue = "#4230"
+
+[[entry]]
 # Crossed the 1000-prod-LoC giant threshold in #3540: the durable pending
 # synthetic-turn-start worker (#3154) accreted the #3174/#3282/#3296/#3303/#3350
 # reconcile-marker chokepoints plus the #3540 no-evict post-abort queue promote.
