+++ src/services/discord/turn_bridge/mod.rs
@@ -22,6 +22,7 @@ mod single_message_footer;
 mod skill_usage;
 mod stale_resume;
 mod status_panel;
+mod stream_loop;
 mod stream_tick;
 mod streaming_edit_text;
 mod task_notification_lifecycle;
@@ -200,10 +201,6 @@ use retry_state::{
     rewind_and_persist_delivery_on_reclaim, sync_response_delivery_state,
     sync_terminal_error_delivery_state_for_bridge_owner,
 };
-use runtime_handoff_loop::{
-    RuntimeHandoffLoopContext, RuntimeHandoffLoopMessage, RuntimeHandoffLoopOutcome,
-    RuntimeHandoffLoopState, handle_runtime_handoff_loop_message,
-};
 use skill_usage::record_skill_usage_from_tool_use;
 use stale_resume::{
     output_file_has_stale_resume_error_after_offset, stream_error_has_stale_resume_error,
@@ -214,7 +211,6 @@ use status_panel::{
     record_status_panel_events, should_open_long_running_placeholder_controller,
     status_panel_completion_ready_after_terminal_body, status_panel_message_id_for_turn,
 };
-use stream_tick::{BridgeStreamTickContext, BridgeStreamTickState, run_bridge_stream_tick};
 use terminal_delivery::{
     BridgeLeaseAcquire, bridge_delivery_lease_for_inflight, bridge_delivery_lease_key_for_inflight,
     bridge_epilogue_clears_inflight, bridge_epilogue_marks_watcher_delivered,
@@ -1346,1359 +1342,101 @@ pub(super) fn spawn_turn_bridge(
             }),
         );
 
-        // #2289 cancel finalization helper. Both the pre-`try_recv` guard
-        // and the post-`try_recv` re-sample funnel through this so the
-        // cancellation bookkeeping (inflight sync, `cancelled = true`,
-        // background-child abort) stays in lock step and cannot drift.
-        // Implemented as a macro so it can mutate locals owned by the
-        // surrounding `while` body without moving them into a closure that
-        // would conflict with `&mut` borrows held elsewhere. The macro
-        // performs the bookkeeping; callers must follow with `break 'outer`
-        // (or be in a position where falling through hits the outer loop
-        // boundary) so the loop exits to the cancel post-processing path.
-        macro_rules! finalize_cancel_inner {
-            () => {{
-                if sync_inflight_restart_mode_from_cancel(
-                    cancel_token.as_ref(),
-                    &mut inflight_state,
-                ) {
-                    let _ = save_inflight_state(&inflight_state);
-                }
-                cancelled = true;
-                close_all_tracked_background_children(
-                    shared_owned.pg_pool.as_ref(),
-                    &mut active_background_child_session_ids,
-                    "aborted",
-                    "turn cancel",
-                )
-                .await;
-            }};
+        let stream_loop_output = stream_loop::run_stream_loop(
+            stream_loop::StreamLoopContext {
+                shared_owned: shared_owned.clone(),
+                gateway: gateway.clone(),
+                channel_id,
+                provider: provider.clone(),
+                cancel_token: cancel_token.clone(),
+                user_text_owned: user_text_owned.clone(),
+                request_owner_name: request_owner_name.clone(),
+                adk_session_key: adk_session_key.clone(),
+                adk_session_name: adk_session_name.clone(),
+                adk_session_info: adk_session_info.clone(),
+                adk_cwd: adk_cwd.clone(),
+                dispatch_id: dispatch_id.clone(),
+                role_binding: role_binding.clone(),
+                turn_id: turn_id.clone(),
+                voice_progress_playback_channel_id,
+                single_message_panel_footer_mode,
+                footer_owner,
+                status_panel_started_at,
+                status_interval,
+                context_window_tokens,
+                context_compact_percent,
+            },
+            stream_loop::StreamLoopState {
+                rx: &mut rx,
+                full_response: &mut full_response,
+                last_edit_text: &mut last_edit_text,
+                done: &mut done,
+                cancelled: &mut cancelled,
+                rx_disconnected: &mut rx_disconnected,
+                current_tool_line: &mut current_tool_line,
+                prev_tool_status: &mut prev_tool_status,
+                last_tool_name: &mut last_tool_name,
+                last_tool_summary: &mut last_tool_summary,
+                accumulated_input_tokens: &mut accumulated_input_tokens,
+                accumulated_cache_create_tokens: &mut accumulated_cache_create_tokens,
+                accumulated_cache_read_tokens: &mut accumulated_cache_read_tokens,
+                accumulated_output_tokens: &mut accumulated_output_tokens,
+                spin_idx: &mut spin_idx,
+                restart_followup_pending: &mut restart_followup_pending,
+                any_tool_used: &mut any_tool_used,
+                has_post_tool_text: &mut has_post_tool_text,
+                tmux_handed_off: &mut tmux_handed_off,
+                watcher_owns_assistant_relay: &mut watcher_owns_assistant_relay,
+                watcher_relay_available_for_turn: &mut watcher_relay_available_for_turn,
+                watcher_handoff_claim_outcome: &mut watcher_handoff_claim_outcome,
+                standby_relay_owns_output: &mut standby_relay_owns_output,
+                last_assistant_text_line: &mut last_assistant_text_line,
+                long_running_placeholder_active: &mut long_running_placeholder_active,
+                active_background_child_session_ids: &mut active_background_child_session_ids,
+                transport_error: &mut transport_error,
+                transcript_events: &mut transcript_events,
+                resume_failure_detected: &mut resume_failure_detected,
+                session_handshake_seen: &mut session_handshake_seen,
+                terminal_session_reset_required: &mut terminal_session_reset_required,
+                recovery_retry: &mut recovery_retry,
+                last_adk_heartbeat: &mut last_adk_heartbeat,
+                pending_stream_messages: &mut pending_stream_messages,
+                pending_status_tool_results: &mut pending_status_tool_results,
+                pending_status_tool_results_by_id: &mut pending_status_tool_results_by_id,
+                last_inflight_long_run_heartbeat: &mut last_inflight_long_run_heartbeat,
+                last_activity_heartbeat_at: &mut last_activity_heartbeat_at,
+                terminal_control_ready_observed: &mut terminal_control_ready_observed,
+                terminal_control_drain_until: &mut terminal_control_drain_until,
+                current_msg_id: &mut current_msg_id,
+                response_sent_offset: &mut response_sent_offset,
+                bridge_confirmed_response_sent_offset: &mut bridge_confirmed_response_sent_offset,
+                streamed_assistant_text_this_turn: &mut streamed_assistant_text_this_turn,
+                streaming_rollover_frozen_msg_ids: &mut streaming_rollover_frozen_msg_ids,
+                terminal_full_replay_cleanup_msg_ids: &mut terminal_full_replay_cleanup_msg_ids,
+                tmux_last_offset: &mut tmux_last_offset,
+                watcher_owner_channel_id: &mut watcher_owner_channel_id,
+                new_session_id: &mut new_session_id,
+                new_raw_provider_session_id: &mut new_raw_provider_session_id,
+                inflight_state: &mut inflight_state,
+                last_status_edit: &mut last_status_edit,
+                first_answer_relayed: &mut first_answer_relayed,
+                last_session_panel_lifecycle_refresh: &mut last_session_panel_lifecycle_refresh,
+                status_panel_msg_id: &mut status_panel_msg_id,
+                last_status_panel_text: &mut last_status_panel_text,
+                status_panel_dirty: &mut status_panel_dirty,
+                last_status_panel_edit: &mut last_status_panel_edit,
+                bridge_spans: &mut bridge_spans,
+                status_panel_generation: &mut status_panel_generation,
+            },
+        )
+        .await;
+        match stream_loop_output.outcome {
+            stream_loop::StreamLoopOutcome::Completed => {}
         }
-
-        let mut pending_long_running_open_after_state_save = None;
-        let mut pending_long_running_retarget_after_state_save = None;
-
-        'outer: while !done
-            || terminal_control_drain_until
-                .is_some_and(|deadline| std::time::Instant::now() < deadline)
-        {
-            let mut state_dirty = false;
-
-            // #2172 cancel boundary: once `done` is true the turn's
-            // terminal outcome (Completed / Done message) has already been
-            // observed; the loop continues only to drain residual control
-            // frames during `terminal_control_drain_until`. A cancel that
-            // arrives in that drain window MUST NOT reclassify the
-            // already-completed turn as cancelled (which would re-run
-            // stop_active_turn and dispatch-cancel finalisation). The
-            // documented "whichever comes first wins" priority is
-            // enforced by gating the cancel arm on `!done`. Cancels
-            // arriving during the drain window break out of the loop
-            // normally as a completed turn.
-            if !done && cancel_requested(Some(cancel_token.as_ref())) {
-                finalize_cancel_inner!();
-                break 'outer;
-            }
-            if done && cancel_requested(Some(cancel_token.as_ref())) {
-                // #2172: cancel-after-Done during terminal drain — exit
-                // the drain immediately as a completed turn instead of
-                // burning the rest of the drain window. Suppressing the
-                // reclassification still preserves the "stop after
-                // completion is a no-op" UX.
-                break 'outer;
-            }
-
-            // #2426 H3/H4 graduation: wait for the next stream frame and
-            // treat the duration as a safety wake, not a pre-drain sleep.
-            // Explicit handoff frames (`TmuxReady` / `ProcessReady` /
-            // `RuntimeReady`) wake this loop immediately and clear
-            // `terminal_control_drain_until` in their handlers.
-            let stream_wait = turn_bridge_stream_wait_duration(
-                done,
-                terminal_control_drain_until,
-                std::time::Instant::now(),
-            );
-            if stream_wait.is_zero() {
-                if let Ok(msg) = rx.try_recv() {
-                    pending_stream_messages.push_back(msg);
-                }
-            } else {
-                match tokio::time::timeout(stream_wait, rx.recv()).await {
-                    Ok(Some(msg)) => pending_stream_messages.push_back(msg),
-                    Ok(None) | Err(_) => {}
-                }
-            }
-
-            if !done && cancel_requested(Some(cancel_token.as_ref())) {
-                finalize_cancel_inner!();
-                break 'outer;
-            }
-            if done && cancel_requested(Some(cancel_token.as_ref())) {
-                // See note above on the cancel-after-Done race.
-                break 'outer;
-            }
-
-            loop {
-                // #2172 cancel boundary: re-check the cancel flag between
-                // drained messages. Without this, the outer loop samples
-                // `cancel_requested` once and then drains EVERY queued
-                // StreamMessage to completion — so a cancel that flips
-                // mid-drain can let a queued `Done` set `done = true`
-                // before the outer cancel-arm runs, which then can no
-                // longer classify the turn as cancelled (the `!done`
-                // gate suppresses it). Break out of the drain on cancel
-                // so the outer cancel-arm gets first claim on the
-                // turn outcome. Frames already pulled before cancel was
-                // observed have been processed (acceptable: they
-                // happened before the user pressed stop); subsequent
-                // frames are left in `rx` and dropped by the bridge
-                // shutdown path.
-                if !done && cancel_requested(Some(cancel_token.as_ref())) {
-                    break;
-                }
-                let next_message = if let Some(msg) = pending_stream_messages.pop_front() {
-                    Ok(msg)
-                } else {
-                    rx.try_recv()
-                };
-                match next_message {
-                    Ok(msg) => {
-                        // #2289 cancel boundary: re-sample `cancel_requested`
-                        // AFTER `try_recv`, but ONLY for variants that flip
-                        // `done = true` (`Done`/`Error`). The pre-recv guard
-                        // samples before the receive; if `/stop` flips the token
-                        // in that gap, letting a terminal arm run sets `done` and
-                        // suppresses the outer cancel arm — recording a completed/
-                        // failed turn the user actually stopped. Drop the frame
-                        // and jump to cancel-finalize.
-                        //
-                        // Scoped to done-setting variants so non-terminal frames
-                        // (`RuntimeReady`, `TmuxReady`, `ProcessReady`,
-                        // `OutputOffset`, `Text`, `RetryBoundary`, …) are still
-                        // processed (they carry handoff paths, offsets, watcher
-                        // debt, session-reset that the cancel path needs); none
-                        // flip `done`, so the next pre-recv cancel guard finalizes
-                        // cancel cleanly. A new terminal variant MUST be added here
-                        // too — see `is_done_setting_terminal_frame`.
-                        if is_done_setting_terminal_frame(&msg)
-                            && should_finalize_cancel_after_recv(
-                                done,
-                                cancel_requested(Some(cancel_token.as_ref())),
-                            )
-                        {
-                            // The dropped frame's bookkeeping (full_response
-                            // resolution, transcript Result/Error,
-                            // placeholder close, transport_error edge)
-                            // is intentionally skipped: the cancel path
-                            // is the authoritative finalizer for the
-                            // turn outcome and runs its own placeholder
-                            // teardown.
-                            finalize_cancel_inner!();
-                            break 'outer;
-                        }
-                        match msg {
-                        StreamMessage::RetryBoundary => {
-                            if provider == ProviderKind::Gemini
-                                && handle_gemini_retry_boundary(
-                                    &mut full_response,
-                                    &mut current_tool_line,
-                                    &mut prev_tool_status,
-                                    &mut last_tool_name,
-                                    &mut last_tool_summary,
-                                    &mut any_tool_used,
-                                    &mut has_post_tool_text,
-                                    &mut response_sent_offset,
-                                    &mut last_edit_text,
-                                    &mut new_session_id,
-                                    &mut new_raw_provider_session_id,
-                                    &mut inflight_state,
-                                )
-                            {
-                                transcript_events.clear();
-                                state_dirty = true;
-                            }
-                        }
-                        StreamMessage::Init {
-                            session_id: sid,
-                            raw_session_id,
-                        } => {
-                            new_session_id = Some(sid.clone());
-                            new_raw_provider_session_id =
-                                raw_session_id.or_else(|| Some(sid.clone()));
-                            inflight_state.session_id = Some(sid);
-                            // #2451 H5: explicit handshake witness — the
-                            // provider has answered with a bound session.
-                            // Any subsequent empty-response classification
-                            // can rely on this instead of elapsed-time
-                            // guessing.
-                            session_handshake_seen = true;
-                            state_dirty = true;
-                        }
-                        StreamMessage::Text { content } => {
-                            let (content, progress_markers) =
-                                if inflight_state.source == crate::dispatch::Source::Voice {
-                                    crate::voice::progress::extract_progress_markers(&content)
-                                } else {
-                                    (content, Vec::new())
-                                };
-                            for marker in progress_markers {
-                                shared_owned.voice_barge_in.publish_progress_for_playback(
-                                    channel_id,
-                                    voice_progress_playback_channel_id,
-                                    marker,
-                                );
-                            }
-                            if content.is_empty() {
-                                continue;
-                            }
-                            streamed_assistant_text_this_turn = true;
-                            // #3608: normalize the chunk boundary so a tool-use
-                            // `\n\n` separator + a chunk that itself begins with
-                            // blank lines does not accumulate into `\n\n\n\n`.
-                            chunk_compose::append_streamed_text_chunk(&mut full_response, &content);
-                            if (watcher_owns_assistant_relay
-                                && watcher_relay_available_for_turn)
-                                || standby_relay_owns_output
-                            {
-                                response_sent_offset = full_response.len();
-                                inflight_state.response_sent_offset = response_sent_offset;
-                            }
-                            // #1255: remember the last non-empty single-line
-                            // assistant prose so we can surface it on a
-                            // long-running tool placeholder card. Mid-stream
-                            // chunks routinely contain newlines, so we walk
-                            // backwards through the lines and pick the most
-                            // recent non-empty one.  `Text` events arrive
-                            // before the immediately-following `ToolUse` event
-                            // in Claude Code's stream ordering, so this
-                            // captures the right hint without buffering.
-                            if let Some(line) = content
-                                .lines()
-                                .filter(|l| !l.trim().is_empty())
-                                .next_back()
-                                .map(str::trim)
-                                .map(str::to_string)
-                            {
-                                last_assistant_text_line = Some(line);
-                            }
-                            push_transcript_event(
-                                &mut transcript_events,
-                                SessionTranscriptEvent {
-                                    kind: SessionTranscriptEventKind::Assistant,
-                                    tool_name: None,
-                                    summary: None,
-                                    content: content.clone(),
-                                    status: Some("success".to_string()),
-                                    is_error: false,
-                                },
-                            );
-                            if any_tool_used {
-                                has_post_tool_text = true;
-                                inflight_state.has_post_tool_text = true;
-                            }
-                            super::formatting::preserve_previous_tool_status(
-                                &mut prev_tool_status,
-                                current_tool_line.as_deref(),
-                                None,
-                            );
-                            current_tool_line = None;
-                            last_tool_name = None;
-                            last_tool_summary = None;
-                            inflight_state.full_response = full_response.clone();
-                            state_dirty = true;
-                        }
-                        StreamMessage::Thinking { summary } => {
-                            let display = thinking_status_line();
-                            status_panel_dirty |= record_status_panel_events(
-                                shared_owned.as_ref(),
-                                channel_id,
-                                vec![StatusEvent::Heartbeat],
-                            );
-                            if inflight_state.source == crate::dispatch::Source::Voice {
-                                shared_owned.voice_barge_in.publish_progress_for_playback(
-                                    channel_id,
-                                    voice_progress_playback_channel_id,
-                                    "thinking",
-                                );
-                            }
-                            // #1113 implicit-terminate: a Thinking event after an
-                            // unfinished ToolUse means the agent moved on without
-                            // emitting a ToolResult. Promote the orphaned tool to
-                            // its terminal (⚠) form before stashing in prev so the
-                            // user does not see a stale ⚙ indicator.
-                            let prev_for_preserve = current_tool_line
-                                .as_deref()
-                                .map(super::formatting::finalize_in_progress_tool_status);
-                            super::formatting::preserve_previous_tool_status(
-                                &mut prev_tool_status,
-                                prev_for_preserve.as_deref(),
-                                Some(display.as_str()),
-                            );
-                            current_tool_line = Some(display);
-                            last_tool_name = None;
-                            last_tool_summary = None;
-                            // Thinking payloads can contain raw model reasoning.
-                            // Keep them out of the user-visible Discord response
-                            // and use only a neutral progress marker.
-                            state_dirty = true;
-                            push_transcript_event(
-                                &mut transcript_events,
-                                redacted_thinking_transcript_event(summary),
-                            );
-                        }
-                        StreamMessage::ToolUse {
-                            name,
-                            input,
-                            tool_use_id,
-                        } => {
-                            // #3084: index the tool name by its tool-use id when
-                            // present so the matching ToolResult resolves the
-                            // exact tool regardless of interleaving; otherwise
-                            // fall back to FIFO ordering.
-                            match tool_use_id.as_deref() {
-                                Some(id) => {
-                                    pending_status_tool_results_by_id
-                                        .insert(id.to_string(), name.clone());
-                                }
-                                None => pending_status_tool_results.push_back(name.clone()),
-                            }
-                            any_tool_used = true;
-                            has_post_tool_text = false;
-                            inflight_state.any_tool_used = true;
-                            inflight_state.has_post_tool_text = false;
-                            if let (Some(pg_pool), Some(session_key)) =
-                                (shared_owned.pg_pool.as_ref(), adk_session_key.as_deref())
-                            {
-                                if let Err(error) =
-                                    mark_session_tool_use_pg(pg_pool, session_key).await
-                                {
-                                    let ts = chrono::Local::now().format("%H:%M:%S");
-                                    tracing::warn!(
-                                        "  [{ts}] ⚠ Failed to update last_tool_at for session {}: {}",
-                                        session_key,
-                                        error
-                                    );
-                                }
-                            }
-                            if shared_owned.pg_pool.is_some() {
-                                match record_skill_usage_from_tool_use(
-                                    shared_owned.pg_pool.as_ref(),
-                                    &name,
-                                    &input,
-                                    adk_session_key.as_deref(),
-                                    role_binding.as_ref(),
-                                )
-                                .await
-                                {
-                                    Ok(Some(_)) => {}
-                                    Ok(None) => {}
-                                    Err(e) => {
-                                        let ts = chrono::Local::now().format("%H:%M:%S");
-                                        tracing::warn!(
-                                            "  [{ts}] ⚠ Failed to record skill usage for tool {}: {}",
-                                            name,
-                                            e
-                                        );
-                                    }
-                                }
-                            }
-                            let summary = format_tool_input(&name, &input);
-                            let display_summary = if summary.trim().is_empty() {
-                                "…".to_string()
-                            } else {
-                                truncate_str(&summary, 120).to_string()
-                            };
-                            let display = format!("⚙ {}: {}", name, display_summary);
-                            if inflight_state.source == crate::dispatch::Source::Voice {
-                                shared_owned.voice_barge_in.publish_progress_for_playback(
-                                    channel_id,
-                                    voice_progress_playback_channel_id,
-                                    format!("tool:{name}:{display_summary}"),
-                                );
-                            }
-                            record_placeholder_live_event(
-                                shared_owned.as_ref(),
-                                channel_id,
-                                super::placeholder_live_events::RecentPlaceholderEvent::tool_use(
-                                    &name, &input,
-                                ),
-                            );
-                            status_panel_dirty |= record_status_panel_events(
-                                shared_owned.as_ref(),
-                                channel_id,
-                                super::placeholder_live_events::status_events_from_tool_use_with_id(
-                                    &name,
-                                    &input,
-                                    tool_use_id.as_deref(),
-                                ),
-                            );
-                            // #1113 implicit-terminate: a new ToolUse arriving
-                            // before the prior ToolResult means the previous
-                            // tool is orphaned (parser miss, parallel-tool
-                            // collapse, or genuine hang upstream). Promote the
-                            // stale ⚙ marker to its terminal ⚠ form before
-                            // stashing in prev_tool_status so the placeholder
-                            // never claims a tool is still running once the
-                            // agent has moved on.
-                            let prev_for_preserve = current_tool_line
-                                .as_deref()
-                                .map(super::formatting::finalize_in_progress_tool_status);
-                            super::formatting::preserve_previous_tool_status(
-                                &mut prev_tool_status,
-                                prev_for_preserve.as_deref(),
-                                Some(display.as_str()),
-                            );
-                            current_tool_line = Some(display.clone());
-                            last_tool_name = Some(name.clone());
-                            last_tool_summary = Some(display_summary.clone());
-                            // #1255 live-turn long-running tool detection.
-                            //
-                            // Classifier returns `Some` for Monitor and for
-                            // Bash/Task/Agent calls with explicit
-                            // `run_in_background=true`.  Everything else
-                            // streams its result inline and never touches the
-                            // placeholder card.
-                            // codex round-11 P2 on PR #1308: restart commands
-                            // already own a planned ♻️ handoff message via the
-                            // `is_dcserver_restart_command(&input)` branch
-                            // below; opening a long-running placeholder on
-                            // the same message_id would let a later
-                            // Done/Result write a generic background card
-                            // over the planned restart notice. Skip setup
-                            // for those.
-                            let long_running_tool =
-                                if !is_dcserver_restart_command(&input) {
-                                    super::formatting::classify_long_running_tool(&name, &input)
-                                } else {
-                                    None
-                                };
-                            if matches!(
-                                long_running_tool,
-                                Some((
-                                    _,
-                                    super::formatting::LongRunningCloseTrigger::BackgroundDispatch,
-                                    _
-                                ))
-                            ) {
-                                if let (Some(pg_pool), Some(parent_session_key)) =
-                                    (shared_owned.pg_pool.as_ref(), adk_session_key.as_deref())
-                                {
-                                    let spawn = BackgroundChildSpawn {
-                                        parent_session_key: parent_session_key.to_string(),
-                                        provider: Some(provider.as_str().to_string()),
-                                        tool_name: name.clone(),
-                                        tool_input: input.clone(),
-                                    };
-                                    match insert_background_child_pg(pg_pool, &spawn).await {
-                                        Ok(Some(child_session_id)) => {
-                                            active_background_child_session_ids
-                                                .push(child_session_id);
-                                        }
-                                        Ok(None) => {}
-                                        Err(error) => {
-                                            let ts = chrono::Local::now().format("%H:%M:%S");
-                                            tracing::warn!(
-                                                "  [{ts}] ⚠ Failed to insert background child session for {}: {}",
-                                                parent_session_key,
-                                                error
-                                            );
-                                        }
-                                    }
-                                }
-                            }
-                            if should_open_long_running_placeholder_controller(
-                                shared_owned.ui.status_panel_v2_enabled,
-                            )
-                                && long_running_placeholder_active.is_none()
-                                && pending_long_running_open_after_state_save.is_none()
-                                && pending_long_running_retarget_after_state_save.is_none()
-                            {
-                                if let Some((reason, close_trigger, reason_detail)) =
-                                    long_running_tool
-                                {
-                                    let started_at_unix = chrono::Utc::now().timestamp();
-                                    let key =
-                                        super::placeholder_controller::PlaceholderKey {
-                                            provider: provider.clone(),
-                                            channel_id,
-                                            message_id: current_msg_id,
-                                        };
-                                    let input_payload =
-                                        super::placeholder_controller::PlaceholderActiveInput {
-                                            reason,
-                                            started_at_unix,
-                                            tool_summary: Some(name.clone()),
-                                            command_summary: Some(display_summary.clone()),
-                                            reason_detail,
-                                            context_line: last_assistant_text_line.clone(),
-                                            request_line: first_request_line(&user_text_owned),
-                                            progress_line: child_progress_line(
-                                                shared_owned.pg_pool.as_ref(),
-                                                adk_session_key.as_deref(),
-                                            )
-                                            .await,
-                                        };
-                                    pending_long_running_open_after_state_save =
-                                        Some((key, input_payload, close_trigger, false));
-                                    inflight_state.long_running_placeholder_active = true;
-                                    state_dirty = true;
-                                }
-                            }
-                            push_transcript_event(
-                                &mut transcript_events,
-                                SessionTranscriptEvent {
-                                    kind: SessionTranscriptEventKind::ToolUse,
-                                    tool_name: Some(name.clone()),
-                                    summary: last_tool_summary.clone(),
-                                    content: input.clone(),
-                                    status: Some("running".to_string()),
-                                    is_error: false,
-                                },
-                            );
-                            if !restart_followup_pending && is_dcserver_restart_command(&input) {
-                                let mut report = RestartCompletionReport::new(
-                                    provider.clone(),
-                                    channel_id.get(),
-                                    "pending",
-                                    format!(
-                                        "dcserver restart requested by `{}`; 새 프로세스가 후속 보고를 이어받을 예정입니다.",
-                                        request_owner_name
-                                    ),
-                                );
-                                report.current_msg_id = Some(current_msg_id.get());
-                                report.channel_name = adk_session_name.clone();
-                                if save_restart_report(&report).is_ok() {
-                                    restart_followup_pending = true;
-                                    inflight_state.set_restart_mode(
-                                        crate::services::discord::InflightRestartMode::DrainRestart,
-                                    );
-                                    let handoff_text = "♻️ dcserver 재시작 중...\n\n재시작 후 현재 turn은 자동 새 턴으로 이어가지 않고, 상태만 다시 확인합니다.";
-                                    let _ = gateway
-                                        .edit_message(channel_id, current_msg_id, handoff_text)
-                                        .await;
-                                    last_edit_text = handoff_text.to_string();
-                                    inflight_state.current_msg_id = current_msg_id.get();
-                                    inflight_state.current_msg_len = handoff_text.len();
-                                    state_dirty = true;
-                                }
-                            }
-                            if !full_response.is_empty() {
-                                // #3608: paragraph separator via the shared
-                                // composition primitive (matched pair with
-                                // `append_streamed_text_chunk`). inflight_state
-                                // / state_dirty stay inline (hot-file #3016:
-                                // only full_response composition is extracted).
-                                chunk_compose::append_tool_boundary_separator(&mut full_response);
-                                inflight_state.full_response = full_response.clone();
-                                state_dirty = true;
-                            }
-                        }
-                        StreamMessage::ToolResult {
-                            content,
-                            is_error,
-                            tool_use_id,
-                        } => {
-                            // #3084: resolve the originating tool by id when the
-                            // result carries one (pairing a delayed subagent
-                            // result to its own ToolUse); otherwise fall back to
-                            // FIFO order for id-less backends.
-                            let status_tool_name = match tool_use_id.as_deref() {
-                                Some(id) => pending_status_tool_results_by_id
-                                    .remove(id)
-                                    .or_else(|| pending_status_tool_results.pop_front()),
-                                None => pending_status_tool_results.pop_front(),
-                            };
-                            if inflight_state.source == crate::dispatch::Source::Voice {
-                                let label = if is_error {
-                                    "tool_result:error"
-                                } else {
-                                    "tool_result:ok"
-                                };
-                                shared_owned.voice_barge_in.publish_progress_for_playback(
-                                    channel_id,
-                                    voice_progress_playback_channel_id,
-                                    label,
-                                );
-                            }
-                            // #1084: flag oversize tool outputs + record metrics.
-                            // Never mutates `content` — the agent and transcript
-                            // still see the raw output; only a warn log + counters
-                            // fire when thresholds are exceeded.
-                            let _ = crate::services::tool_output_guard::observe(
-                                last_tool_name.as_deref(),
-                                is_error,
-                                &content,
-                            );
-                            if is_error {
-                                record_placeholder_live_event(
-                                    &shared_owned,
-                                    channel_id,
-                                    super::placeholder_live_events::RecentPlaceholderEvent::tool_error(
-                                        &content,
-                                    ),
-                                );
-                            }
-                            status_panel_dirty |= record_status_panel_events(
-                                shared_owned.as_ref(),
-                                channel_id,
-                                super::placeholder_live_events::status_events_from_tool_result_with_id(
-                                    status_tool_name.as_deref(),
-                                    is_error,
-                                    tool_use_id.as_deref(),
-                                ),
-                            );
-                            // #1255: a long-running tool's ToolResult can move
-                            // the background card terminal while keeping the
-                            // placeholder visible for the rest of the turn; the
-                            // controller remains idempotent. codex round-2 P1:
-                            // only Monitor-style ToolResults are real
-                            // completions. Background Bash/Task/Agent results
-                            // are job/task acks, so keep those open until
-                            // Done/cancel.
-                            if let Some((key, snapshot, close_trigger, ack_consumed)) =
-                                long_running_placeholder_active.take()
-                            {
-                                let monitor_like = matches!(
-                                    close_trigger,
-                                    super::formatting::LongRunningCloseTrigger::MonitorLike
-                                );
-                                // codex round-6 P2: only the FIRST ToolResult
-                                // after the background ToolUse can be its
-                                // dispatch ack. Subsequent ToolResults belong
-                                // to other foreground tools and must not close
-                                // the still-running background card. Once the
-                                // ack is consumed, re-stash unconditionally
-                                // until `Done`/cancel.
-                                let is_dispatch_ack =
-                                    !monitor_like && !ack_consumed;
-                                // codex round-5 P2: an `is_error` ToolResult on
-                                // the background dispatch ack (launch failure)
-                                // closes the card as Aborted; otherwise `Done`
-                                // would later mark it Completed and Discord
-                                // would report a failed background launch as ✅.
-                                if monitor_like || (is_dispatch_ack && is_error) {
-                                    if is_dispatch_ack && is_error {
-                                        close_next_tracked_background_child(
-                                            shared_owned.pg_pool.as_ref(),
-                                            &mut active_background_child_session_ids,
-                                            "aborted",
-                                            "failed background dispatch ack",
-                                        )
-                                        .await;
-                                    }
-                                    let target = if is_error {
-                                        super::placeholder_controller::PlaceholderLifecycle::Aborted
-                                    } else {
-                                        super::placeholder_controller::PlaceholderLifecycle::Completed
-                                    };
-                                    let pending_retarget_matches_key =
-                                        pending_long_running_retarget_after_state_save
-                                            .as_ref()
-                                            .is_some_and(|(pending_key, _, _, _, _)| {
-                                                *pending_key == key
-                                            });
-                                    if pending_retarget_matches_key {
-                                        let _ =
-                                            pending_long_running_retarget_after_state_save.take();
-                                        shared_owned.ui.placeholder_controller.detach(&key);
-                                        inflight_state.long_running_placeholder_active = false;
-                                        state_dirty = true;
-                                    } else {
-                                        let outcome = shared_owned
-                                            .ui.placeholder_controller
-                                            .transition(gateway.as_ref(), key.clone(), target)
-                                            .await;
-                                        // codex round-10 P2: only clear flag on
-                                        // committed/already-terminal outcome.
-                                        use super::placeholder_controller::PlaceholderControllerOutcome::*;
-                                        if matches!(outcome, Edited | Coalesced | AlreadyTerminal)
-                                        {
-                                            inflight_state.long_running_placeholder_active = false;
-                                            state_dirty = true;
-                                        } else {
-                                            // EditFailed — keep the placeholder
-                                            // active so the next event/sweeper
-                                            // can retry the terminal edit.
-                                            long_running_placeholder_active =
-                                                Some((key, snapshot, close_trigger, ack_consumed));
-                                        }
-                                    }
-                                } else {
-                                    // Successful background dispatch ack OR a
-                                    // later unrelated ToolResult — re-stash so
-                                    // `Done`/cancel can close the card. Mark
-                                    // the ack consumed so future is_error
-                                    // results from other tools don't abort us.
-                                    let active_key_for_pending_update = key.clone();
-                                    long_running_placeholder_active = Some((
-                                        key,
-                                        snapshot,
-                                        close_trigger,
-                                        true,
-                                    ));
-                                    if let Some((
-                                        pending_key,
-                                        _,
-                                        _,
-                                        pending_ack_consumed,
-                                        _,
-                                    )) = pending_long_running_retarget_after_state_save.as_mut()
-                                        && *pending_key == active_key_for_pending_update
-                                    {
-                                        *pending_ack_consumed = true;
-                                    }
-                                }
-                            } else if let Some((_, _, close_trigger, ack_consumed)) =
-                                pending_long_running_open_after_state_save.as_mut()
-                            {
-                                let monitor_like = matches!(
-                                    *close_trigger,
-                                    super::formatting::LongRunningCloseTrigger::MonitorLike
-                                );
-                                let is_dispatch_ack = !monitor_like && !*ack_consumed;
-                                if monitor_like || (is_dispatch_ack && is_error) {
-                                    pending_long_running_open_after_state_save = None;
-                                    inflight_state.long_running_placeholder_active = false;
-                                    state_dirty = true;
-                                } else {
-                                    *ack_consumed = true;
-                                }
-                            }
-                            // Reset the assistant-line summary so the next
-                            // long-running tool call captures its own context.
-                            last_assistant_text_line = None;
-                            if let Some(ref tn) = last_tool_name {
-                                let status = if is_error { "✗" } else { "✓" };
-                                let detail = last_tool_summary
-                                    .as_deref()
-                                    .filter(|s| !s.is_empty() && *s != "…")
-                                    .map(|s| format!("{} {}: {}", status, tn, s))
-                                    .unwrap_or_else(|| format!("{} {}", status, tn));
-                                super::formatting::preserve_previous_tool_status(
-                                    &mut prev_tool_status,
-                                    current_tool_line.as_deref(),
-                                    Some(detail.as_str()),
-                                );
-                                current_tool_line = Some(detail);
-                            }
-                            push_transcript_event(
-                                &mut transcript_events,
-                                SessionTranscriptEvent {
-                                    kind: if is_error {
-                                        SessionTranscriptEventKind::Error
-                                    } else {
-                                        SessionTranscriptEventKind::ToolResult
-                                    },
-                                    tool_name: last_tool_name.clone(),
-                                    summary: last_tool_summary.clone(),
-                                    content,
-                                    status: Some(
-                                        if is_error { "error" } else { "success" }.to_string(),
-                                    ),
-                                    is_error,
-                                },
-                            );
-                        }
-                        StreamMessage::TaskNotification { tool_use_id, summary, status, kind, .. } => {
-                            inflight_state.task_notification_kind =
-                                merge_task_notification_kind(inflight_state.task_notification_kind, kind);
-                            state_dirty = true;
-                            record_placeholder_live_event(
-                                shared_owned.as_ref(),
-                                channel_id,
-                                super::placeholder_live_events::RecentPlaceholderEvent::task_notification(
-                                    kind.as_str(),
-                                    &status,
-                                    &summary,
-                                ),
-                            );
-                            status_panel_dirty |= record_status_panel_events(
-                                shared_owned.as_ref(),
-                                channel_id,
-                                super::placeholder_live_events::status_events_from_task_notification_with_tool_use_id(
-                                    kind.as_str(),
-                                    &status,
-                                    &summary,
-                                    tool_use_id.as_deref(),
-                                ),
-                            );
-                            if single_message_panel_footer_mode {
-                                let indicator =
-                                    super::single_message_panel::single_message_panel_spinner_frame(
-                                        spin_idx,
-                                    );
-                                spin_idx = spin_idx.wrapping_add(1);
-                                refresh_bridge_footer(
-                                    shared_owned.as_ref(),
-                                    channel_id,
-                                    footer_owner,
-                                    indicator,
-                                )
-                                .await;
-                            }
-                            if task_notification_closes_background_child(kind, &status) {
-                                let close_status = if matches!(
-                                    status.trim().to_ascii_lowercase().as_str(),
-                                    "aborted" | "cancelled" | "canceled" | "failed" | "error"
-                                ) {
-                                    "aborted"
-                                } else {
-                                    "completed"
-                                };
-                                close_next_tracked_background_child(
-                                    shared_owned.pg_pool.as_ref(),
-                                    &mut active_background_child_session_ids,
-                                    close_status,
-                                    "task notification",
-                                )
-                                .await;
-                                // #1670: `merge_task_notification_kind` is an
-                                // absorb operator (priority-max). Without an
-                                // explicit release on the terminal status the
-                                // outer `inflight_state.task_notification_kind`
-                                // sticks at Subagent/Background past the child
-                                // close, which then misroutes downstream
-                                // suppression decisions and persists into the
-                                // saved inflight when the watcher takes over.
-                                //
-                                // codex P2 followup: only release the closed
-                                // child's kind once ALL tracked children have
-                                // closed. If a lower-priority child is the
-                                // last tracked one while a higher-priority
-                                // active classification is absorbed, keep the
-                                // higher-priority kind instead of clearing it.
-                                if active_background_child_session_ids.is_empty() {
-                                    inflight_state.task_notification_kind =
-                                        release_task_notification_kind(
-                                            inflight_state.task_notification_kind,
-                                            kind,
-                                        );
-                                }
-                            }
-                            push_transcript_event(
-                                &mut transcript_events,
-                                SessionTranscriptEvent {
-                                    kind: SessionTranscriptEventKind::Task,
-                                    tool_name: None,
-                                    summary: Some(summary.clone()),
-                                    content: summary,
-                                    status: Some("info".to_string()),
-                                    is_error: false,
-                                },
-                            );
-                        }
-                        StreamMessage::Done {
-                            result,
-                            session_id: sid,
-                        } => {
-                            let session_died_retry = result == "__session_died_retry__";
-                            if session_died_retry {
-                                // Recovery reader requests the generic Discord-history
-                                // auto-retry when the resumed session dies pre-completion.
-                                recovery_retry = true;
-                            }
-                            if pending_long_running_open_after_state_save.take().is_some() {
-                                inflight_state.long_running_placeholder_active = false;
-                                let _ = save_inflight_state(&inflight_state);
-                            }
-                            // #1255: turn finished while a long-running placeholder
-                            // is still Active — close it now so the user does not
-                            // stare at a stale 🔄 card. Idempotent if a prior
-                            // ToolResult already fired Completed.
-                            if let Some((key, snapshot, close_trigger, ack_consumed)) =
-                                long_running_placeholder_active.take()
-                            {
-                                let target = if session_died_retry {
-                                    super::placeholder_controller::PlaceholderLifecycle::Aborted
-                                } else {
-                                    super::placeholder_controller::PlaceholderLifecycle::Completed
-                                };
-                                let pending_retarget_matches_key =
-                                    pending_long_running_retarget_after_state_save
-                                        .as_ref()
-                                        .is_some_and(|(pending_key, _, _, _, _)| {
-                                            *pending_key == key
-                                        });
-                                if pending_retarget_matches_key {
-                                    let _ = pending_long_running_retarget_after_state_save.take();
-                                    shared_owned.ui.placeholder_controller.detach(&key);
-                                    inflight_state.long_running_placeholder_active = false;
-                                } else {
-                                    let outcome = shared_owned
-                                        .ui.placeholder_controller
-                                        .transition(gateway.as_ref(), key.clone(), target)
-                                        .await;
-                                    // codex round-10/11 P2/P3: on `EditFailed`,
-                                    // re-stash the tuple so subsequent
-                                    // streaming/sweeper paths can retry the
-                                    // terminal edit. Only clear the persisted
-                                    // flag on a committed (or already-terminal)
-                                    // transition.
-                                    use super::placeholder_controller::PlaceholderControllerOutcome::*;
-                                    if matches!(outcome, Edited | Coalesced | AlreadyTerminal) {
-                                        inflight_state.long_running_placeholder_active = false;
-                                    } else {
-                                        long_running_placeholder_active =
-                                            Some((key, snapshot, close_trigger, ack_consumed));
-                                    }
-                                }
-                            }
-                            if let Some(resolved) = resolve_done_response(
-                                &full_response,
-                                &result,
-                                any_tool_used,
-                                has_post_tool_text,
-                            ) {
-                                full_response = resolved;
-                                // #3419 R3 (codex MEDIUM): a short sentinel/tool-only
-                                // resolved body shrinks full_response below a prior
-                                // streamed offset that the >8000-byte replay gate
-                                // (`done_result_requires_full_terminal_replay`) does not
-                                // reset → out of bounds (watcher empty-slice wedge).
-                                // `sync_response_delivery_state` clamps to len AND walks
-                                // back to a valid char boundary (a bare `.min(len())`
-                                // could land mid multibyte char) and mirrors both.
-                                sync_response_delivery_state(
-                                    &full_response,
-                                    &mut response_sent_offset,
-                                    &mut inflight_state,
-                                );
-                            }
-                            if done_result_requires_full_terminal_replay(
-                                &full_response,
-                                &result,
-                                response_sent_offset,
-                                streamed_assistant_text_this_turn,
-                            ) {
-                                tracing::info!(
-                                    target: "agentdesk::codex_rollout_handoff",
-                                    provider = %provider.as_str(),
-                                    channel = channel_id.get(),
-                                    previous_response_sent_offset = response_sent_offset,
-                                    full_response_len = full_response.len(),
-                                    done_result_len = result.len(),
-                                    frozen_rollover_messages = streaming_rollover_frozen_msg_ids.len(),
-                                    "turn_bridge reset terminal delivery offset for authoritative Done body"
-                                );
-                                terminal_full_replay_cleanup_msg_ids =
-                                    streaming_rollover_frozen_msg_ids.clone();
-                                response_sent_offset = 0;
-                            }
-                            if let Some(s) = sid {
-                                new_session_id = Some(s.clone());
-                                inflight_state.session_id = Some(s);
-                            }
-                            if !session_died_retry {
-                                push_transcript_event(
-                                    &mut transcript_events,
-                                    SessionTranscriptEvent {
-                                        kind: SessionTranscriptEventKind::Result,
-                                        tool_name: None,
-                                        summary: Some(if result.trim().is_empty() {
-                                            "Turn completed".to_string()
-                                        } else {
-                                            truncate_str(&result, 120).to_string()
-                                        }),
-                                        content: result,
-                                        status: Some("success".to_string()),
-                                        is_error: false,
-                                    },
-                                );
-                            }
-                            if session_died_retry {
-                                close_all_tracked_background_children(
-                                    shared_owned.pg_pool.as_ref(),
-                                    &mut active_background_child_session_ids,
-                                    "aborted",
-                                    "turn done",
-                                )
-                                .await;
-                            }
-                            state_dirty = true;
-                            done = true;
-                            // #2449: only arm the 250ms drain when handoff is still
-                            // ambiguous; fresh managed turns may pre-stamp runtime_kind.
-                            if !terminal_control_ready_observed {
-                                terminal_control_drain_until = Some(
-                                    std::time::Instant::now()
-                                        + std::time::Duration::from_millis(250),
-                                );
-                            }
-                        }
-                        StreamMessage::Error {
-                            message, stderr, ..
-                        } => {
-                            let is_stale_resume =
-                                stream_error_has_stale_resume_error(&message, &stderr);
-                            let session_reset_required =
-                                stream_error_requires_terminal_session_reset(&message, &stderr);
-                            transport_error = true;
-                            let combined = format!("{} {}", message, stderr).to_lowercase();
-                            if combined.contains("prompt is too long")
-                                || combined.contains("prompt too long")
-                                || combined.contains("context_length_exceeded")
-                                || combined.contains("max_tokens")
-                                || combined.contains("context window")
-                                || combined.contains("token limit")
-                            {
-                                // Prompt too long is not a terminal failure — user can retry
-                                // with a shorter message or /compact. Don't mark as transport error.
-                                transport_error = false;
-                                full_response = "⚠️ __prompt too long__".to_string();
-                            } else if is_stale_resume {
-                                // Recoverable stale resume: auto-retry with a fresh provider
-                                // session instead of failing the current dispatch/turn.
-                                transport_error = false;
-                                resume_failure_detected = true;
-                                if !stderr.is_empty() {
-                                    full_response = format!(
-                                        "Error: {}\nstderr: {}",
-                                        message,
-                                        truncate_str(&stderr, 500)
-                                    );
-                                } else {
-                                    full_response = format!("Error: {}", message);
-                                }
-                            } else if !stderr.is_empty() {
-                                full_response = format!(
-                                    "Error: {}\nstderr: {}",
-                                    message,
-                                    truncate_str(&stderr, 500)
-                                );
-                            } else {
-                                full_response = format!("Error: {}", message);
-                            }
-                            sync_terminal_error_delivery_state_for_bridge_owner(
-                                &full_response,
-                                &mut response_sent_offset,
-                                &mut bridge_confirmed_response_sent_offset,
-                                &mut inflight_state,
-                                channel_id,
-                                watcher_owns_assistant_relay && watcher_relay_available_for_turn,
-                            );
-                            push_transcript_event(
-                                &mut transcript_events,
-                                SessionTranscriptEvent {
-                                    kind: SessionTranscriptEventKind::Error,
-                                    tool_name: last_tool_name.clone(),
-                                    summary: Some(message.clone()),
-                                    content: if stderr.trim().is_empty() {
-                                        message.clone()
-                                    } else {
-                                        format!("{message}\n{stderr}")
-                                    },
-                                    status: Some("error".to_string()),
-                                    is_error: true,
-                                },
-                            );
-                            if session_reset_required {
-                                terminal_session_reset_required = true;
-                                clear_local_session_state(
-                                    &mut new_session_id,
-                                    &mut new_raw_provider_session_id,
-                                    &mut inflight_state,
-                                );
-                                let ts = chrono::Local::now().format("%H:%M:%S");
-                                tracing::warn!(
-                                    "  [{ts}] ⚠ Clearing stored provider session after terminal {} session failure (channel {})",
-                                    provider.as_str(),
-                                    channel_id,
-                                );
-                            }
-                            close_all_tracked_background_children(
-                                shared_owned.pg_pool.as_ref(),
-                                &mut active_background_child_session_ids,
-                                "aborted",
-                                "stream error",
-                            )
-                            .await;
-                            state_dirty = true;
-                            done = true;
-                            terminal_control_drain_until = None;
-                        }
-                        StreamMessage::StatusUpdate {
-                            input_tokens,
-                            cache_create_tokens,
-                            cache_read_tokens,
-                            output_tokens,
-                            ..
-                        } => {
-                            let has_context_token_data = input_tokens.is_some()
-                                || cache_create_tokens.is_some()
-                                || cache_read_tokens.is_some();
-                            // Token fields are provider-normalized snapshots,
-                            // not deltas. Claude reports uncached input,
-                            // cache writes, and cache reads separately, while
-                            // other providers may omit unavailable cache
-                            // fields. Keep the largest context-occupancy
-                            // snapshot seen in this turn so late partial
-                            // status events cannot make the live panel shrink.
-                            if has_context_token_data {
-                                apply_context_token_update(
-                                    &mut accumulated_input_tokens,
-                                    &mut accumulated_cache_create_tokens,
-                                    &mut accumulated_cache_read_tokens,
-                                    input_tokens,
-                                    cache_create_tokens,
-                                    cache_read_tokens,
-                                );
-                            }
-                            if let Some(ot) = output_tokens {
-                                accumulated_output_tokens = accumulated_output_tokens.max(ot);
-                            }
-                            if shared_owned.ui.status_panel_v2_enabled && has_context_token_data {
-                                let context_provider_session_id = new_raw_provider_session_id
-                                    .as_deref()
-                                    .or(new_session_id.as_deref())
-                                    .or(inflight_state.session_id.as_deref());
-                                let context_dirty = shared_owned
-                                    .ui.placeholder_live_events
-                                    .set_context_panel_usage(
-                                        channel_id,
-                                        context_provider_session_id,
-                                        accumulated_input_tokens,
-                                        accumulated_cache_create_tokens,
-                                        accumulated_cache_read_tokens,
-                                        context_window_tokens,
-                                        context_compact_percent,
-                                    );
-                                status_panel_dirty |= context_dirty;
-                            }
-                        }
-                        StreamMessage::StatusEvents { events } => {
-                            status_panel_dirty |= record_status_panel_events(
-                                shared_owned.as_ref(),
-                                channel_id,
-                                events,
-                            );
-                        }
-                        StreamMessage::TmuxReady {
-                            output_path,
-                            input_fifo_path,
-                            tmux_session_name,
-                            last_offset,
-                        } => {
-                            let outcome = handle_runtime_handoff_loop_message(
-                                RuntimeHandoffLoopMessage::TmuxReady {
-                                    output_path,
-                                    input_fifo_path,
-                                    tmux_session_name,
-                                    last_offset,
-                                },
-                                RuntimeHandoffLoopContext {
-                                    shared_owned: &shared_owned,
-                                    provider: &provider,
-                                    channel_id,
-                                    done,
-                                    adk_session_name: &adk_session_name,
-                                },
-                                RuntimeHandoffLoopState {
-                                    terminal_control_ready_observed: &mut terminal_control_ready_observed,
-                                    tmux_last_offset: &mut tmux_last_offset,
-                                    inflight_state: &mut inflight_state,
-                                    watcher_owner_channel_id: &mut watcher_owner_channel_id,
-                                    standby_relay_owns_output: &mut standby_relay_owns_output,
-                                    watcher_relay_available_for_turn: &mut watcher_relay_available_for_turn,
-                                    watcher_handoff_claim_outcome: &mut watcher_handoff_claim_outcome,
-                                    tmux_handed_off: &mut tmux_handed_off,
-                                    watcher_owns_assistant_relay: &mut watcher_owns_assistant_relay,
-                                    state_dirty: &mut state_dirty,
-                                    terminal_control_drain_until: &mut terminal_control_drain_until,
-                                    last_activity_heartbeat_at: &mut last_activity_heartbeat_at,
-                                },
-                            )
-                            .await;
-                            match outcome {
-                                RuntimeHandoffLoopOutcome::ContinueDraining => {}
-                            }
-                        }
-                        StreamMessage::RuntimeReady { handoff } => {
-                            let outcome = handle_runtime_handoff_loop_message(
-                                RuntimeHandoffLoopMessage::RuntimeReady { handoff },
-                                RuntimeHandoffLoopContext {
-                                    shared_owned: &shared_owned,
-                                    provider: &provider,
-                                    channel_id,
-                                    done,
-                                    adk_session_name: &adk_session_name,
-                                },
-                                RuntimeHandoffLoopState {
-                                    terminal_control_ready_observed: &mut terminal_control_ready_observed,
-                                    tmux_last_offset: &mut tmux_last_offset,
-                                    inflight_state: &mut inflight_state,
-                                    watcher_owner_channel_id: &mut watcher_owner_channel_id,
-                                    standby_relay_owns_output: &mut standby_relay_owns_output,
-                                    watcher_relay_available_for_turn: &mut watcher_relay_available_for_turn,
-                                    watcher_handoff_claim_outcome: &mut watcher_handoff_claim_outcome,
-                                    tmux_handed_off: &mut tmux_handed_off,
-                                    watcher_owns_assistant_relay: &mut watcher_owns_assistant_relay,
-                                    state_dirty: &mut state_dirty,
-                                    terminal_control_drain_until: &mut terminal_control_drain_until,
-                                    last_activity_heartbeat_at: &mut last_activity_heartbeat_at,
-                                },
-                            )
-                            .await;
-                            match outcome {
-                                RuntimeHandoffLoopOutcome::ContinueDraining => {}
-                            }
-                        }
-                        StreamMessage::ProcessReady {
-                            output_path,
-                            session_name,
-                            last_offset,
-                        } => {
-                            let outcome = handle_runtime_handoff_loop_message(
-                                RuntimeHandoffLoopMessage::ProcessReady {
-                                    output_path,
-                                    session_name,
-                                    last_offset,
-                                },
-                                RuntimeHandoffLoopContext {
-                                    shared_owned: &shared_owned,
-                                    provider: &provider,
-                                    channel_id,
-                                    done,
-                                    adk_session_name: &adk_session_name,
-                                },
-                                RuntimeHandoffLoopState {
-                                    terminal_control_ready_observed: &mut terminal_control_ready_observed,
-                                    tmux_last_offset: &mut tmux_last_offset,
-                                    inflight_state: &mut inflight_state,
-                                    watcher_owner_channel_id: &mut watcher_owner_channel_id,
-                                    standby_relay_owns_output: &mut standby_relay_owns_output,
-                                    watcher_relay_available_for_turn: &mut watcher_relay_available_for_turn,
-                                    watcher_handoff_claim_outcome: &mut watcher_handoff_claim_outcome,
-                                    tmux_handed_off: &mut tmux_handed_off,
-                                    watcher_owns_assistant_relay: &mut watcher_owns_assistant_relay,
-                                    state_dirty: &mut state_dirty,
-                                    terminal_control_drain_until: &mut terminal_control_drain_until,
-                                    last_activity_heartbeat_at: &mut last_activity_heartbeat_at,
-                                },
-                            )
-                            .await;
-                            match outcome {
-                                RuntimeHandoffLoopOutcome::ContinueDraining => {}
-                            }
-                        }
-                        StreamMessage::OutputOffset { offset } => {
-                            let outcome = handle_runtime_handoff_loop_message(
-                                RuntimeHandoffLoopMessage::OutputOffset { offset },
-                                RuntimeHandoffLoopContext {
-                                    shared_owned: &shared_owned,
-                                    provider: &provider,
-                                    channel_id,
-                                    done,
-                                    adk_session_name: &adk_session_name,
-                                },
-                                RuntimeHandoffLoopState {
-                                    terminal_control_ready_observed: &mut terminal_control_ready_observed,
-                                    tmux_last_offset: &mut tmux_last_offset,
-                                    inflight_state: &mut inflight_state,
-                                    watcher_owner_channel_id: &mut watcher_owner_channel_id,
-                                    standby_relay_owns_output: &mut standby_relay_owns_output,
-                                    watcher_relay_available_for_turn: &mut watcher_relay_available_for_turn,
-                                    watcher_handoff_claim_outcome: &mut watcher_handoff_claim_outcome,
-                                    tmux_handed_off: &mut tmux_handed_off,
-                                    watcher_owns_assistant_relay: &mut watcher_owns_assistant_relay,
-                                    state_dirty: &mut state_dirty,
-                                    terminal_control_drain_until: &mut terminal_control_drain_until,
-                                    last_activity_heartbeat_at: &mut last_activity_heartbeat_at,
-                                },
-                            )
-                            .await;
-                            match outcome {
-                                RuntimeHandoffLoopOutcome::ContinueDraining => {}
-                            }
-                        }
-                        }
-                    }
-                    Err(tokio::sync::mpsc::error::TryRecvError::Empty) => break,
-                    Err(tokio::sync::mpsc::error::TryRecvError::Disconnected) => {
-                        // #2289 cancel boundary: re-sample cancel AFTER the
-                        // receiver reports disconnect. If `/stop` flipped
-                        // the token between the pre-recv guard and this
-                        // arm, letting the disconnect set `done = true`
-                        // and exit the inner loop would cause the outer
-                        // cancel arm (gated on `!done`) to skip
-                        // finalisation, leaving the turn recorded as
-                        // completed/empty instead of stopped.
-                        if should_finalize_cancel_after_recv(
-                            done,
-                            cancel_requested(Some(cancel_token.as_ref())),
-                        ) {
-                            finalize_cancel_inner!();
-                            break 'outer;
-                        }
-                        rx_disconnected = true;
-                        done = true;
-                        terminal_control_drain_until = None;
-                        break;
-                    }
-                }
-            }
-
-            run_bridge_stream_tick(
-                BridgeStreamTickContext {
-                    shared_owned: shared_owned.clone(),
-                    gateway: gateway.clone(),
-                    channel_id,
-                    provider: &provider,
-                    turn_id: turn_id.as_str(),
-                    status_interval,
-                    single_message_panel_footer_mode,
-                    footer_owner,
-                    status_panel_started_at,
-                    done,
-                    standby_relay_owns_output,
-                    watcher_owner_channel_id,
-                    full_response: full_response.as_str(),
-                    dispatch_id: dispatch_id.clone(),
-                    adk_session_key: adk_session_key.clone(),
-                    adk_session_name: adk_session_name.clone(),
-                    adk_session_info: adk_session_info.clone(),
-                    adk_cwd: adk_cwd.clone(),
-                    role_binding: role_binding.clone(),
-                    current_tool_line: current_tool_line.clone(),
-                    last_tool_name: last_tool_name.clone(),
-                    last_tool_summary: last_tool_summary.clone(),
-                    prev_tool_status: prev_tool_status.clone(),
-                    spinner: SPINNER,
-                    live_long_run_heartbeat_interval: LIVE_LONG_RUN_HEARTBEAT_INTERVAL,
-                },
-                BridgeStreamTickState {
-                    state_dirty: &mut state_dirty,
-                    last_session_panel_lifecycle_refresh: &mut last_session_panel_lifecycle_refresh,
-                    status_panel_dirty: &mut status_panel_dirty,
-                    spin_idx: &mut spin_idx,
-                    last_status_panel_edit: &mut last_status_panel_edit,
-                    last_status_edit: &mut last_status_edit,
-                    status_panel_msg_id: &mut status_panel_msg_id,
-                    last_status_panel_text: &mut last_status_panel_text,
-                    watcher_owns_assistant_relay: &mut watcher_owns_assistant_relay,
-                    watcher_relay_available_for_turn: &mut watcher_relay_available_for_turn,
-                    response_sent_offset: &mut response_sent_offset,
-                    bridge_confirmed_response_sent_offset: &mut bridge_confirmed_response_sent_offset,
-                    streaming_rollover_frozen_msg_ids: &mut streaming_rollover_frozen_msg_ids,
-                    current_msg_id: &mut current_msg_id,
-                    last_edit_text: &mut last_edit_text,
-                    first_answer_relayed: &mut first_answer_relayed,
-                    inflight_state: &mut inflight_state,
-                    bridge_spans: &mut bridge_spans,
-                    status_panel_generation: &mut status_panel_generation,
-                    pending_long_running_open_after_state_save: &mut pending_long_running_open_after_state_save,
-                    pending_long_running_retarget_after_state_save: &mut pending_long_running_retarget_after_state_save,
-                    long_running_placeholder_active: &mut long_running_placeholder_active,
-                    last_adk_heartbeat: &mut last_adk_heartbeat,
-                    last_inflight_long_run_heartbeat: &mut last_inflight_long_run_heartbeat,
-                },
-            )
-            .await;
-
-        }
-
-        // #3813 AC#1 tail: emit bridge-side latency spans once at loop exit
-        // (observation-only; self-suppresses when no bridge relay happened).
-        bridge_spans.log(channel_id.get(), provider.as_str());
+        let pending_long_running_open_after_state_save =
+            stream_loop_output.pending_long_running_open_after_state_save;
+        let pending_long_running_retarget_after_state_save =
+            stream_loop_output.pending_long_running_retarget_after_state_save;
 
         let post_loop_finalize_output = post_loop_finalize::run_post_loop_finalize(
             post_loop_finalize::PostLoopFinalizeContext {
